scrypt_doge_refactored/requirements.txt
# Core dependencies
numpy>=1.21.0
psutil>=5.8.0
aiohttp>=3.7.4
asyncio>=3.4.3
aiofiles>=0.8.0

# Web framework
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=1.8.0

# Configuration management
PyYAML>=5.4.1
python-decouple>=3.4

# Security
cryptography>=3.4.7

# Monitoring
prometheus-client>=0.11.0

# Hardware integration
pyasic>=0.4.0

# Database
pymongo>=4.0.0
dnspython>=2.0.0

# Dependency injection
dependency-injector>=4.40.0

# Testing
pytest>=6.2.4
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1

# Code quality
ruff>=0.0.272